+++ CHANGELOG.md
@@ -38,8 +38,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - configure open-code synchronization workflow
 - add mapping connector and integrate into ifsg, seq-repo and sumo
-- add odk extract
+- add odk extract and transform
 - LDAP and Organigram integration for seq-repo
+
 
 ### Changes